piska/coact_client/requirements.txt
# Production dependencies
aiohttp>=3.9.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pillow>=10.0.0
pyautogui>=0.9.54
psutil>=5.9.0
cryptography>=41.0.0
click>=8.1.0
rich>=13.0.0
tenacity>=8.2.0
python-multipart>=0.0.6
platformdirs>=4.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0